# contracts/typescript/api-types.ts
/*
 * Generated from contracts/openapi.json.
 * Do not edit this file manually.
 */

/**
 * Request containing one practical coupon-analysis document.
 */
export interface CouponAnalysisRunRequest {
  analysis_document: Record<string, unknown>;
}

/**
 * Versioned practical coupon-analysis result envelope.
 */
export interface CouponAnalysisRunResponse {
  result: Record<string, unknown>;
}

/**
 * Response containing available coupon game types.
 */
export interface CouponListResponse {
  count: number;
  game_types: Array<string>;
}

/**
 * Metadata describing one imported coupon.
 */
export interface CouponMetadataResponse {
  deadline: string | null;
  expected_match_count?: number | null;
  game_type: string;
  game_type_display: string;
  id: string | null;
  imported_at: string;
  match_count: number;
  source: string;
  source_display: string;
}

/**
 * Request containing analysis and reduction configuration data.
 */
export interface CouponReductionRunRequest {
  analysis_document: Record<string, unknown>;
  reduction_configuration: Record<string, unknown>;
}

/**
 * Versioned practical coupon-reduction result envelope.
 */
export interface CouponReductionRunResponse {
  result: Record<string, unknown>;
}

/**
 * Complete versioned coupon response.
 */
export interface CouponResponse {
  coupon: CouponMetadataResponse;
  matches: Array<MatchResponse>;
  schema_version: string;
}

/**
 * Standard API error response.
 */
export interface ErrorResponse {
  detail: string;
}

export interface HTTPValidationError {
  detail?: Array<ValidationError>;
}

/**
 * Response returned by the health endpoint.
 */
export interface HealthResponse {
  service: string;
  status: string;
  version: string;
}

/**
 * API representation of one football match.
 */
export interface MatchResponse {
  away_team: string;
  competition: string | null;
  home_team: string;
  kickoff: string | null;
  number: number;
  status: string;
}

export interface ValidationError {
  ctx?: Record<string, unknown>;
  input?: unknown;
  loc: Array<string | number>;
  msg: string;
  type: string;
}